Transferring classes to separate files. Structuring of components
