+++ lab_06/.gitignore
@@ -1,0 +1,3 @@
+bin
+obj
+.vs
