+++ lab_06/ConsoleApp1/Commands/ICommand.cs
@@ -1,0 +1,8 @@
+﻿// src/ICommand.cs
+namespace ATMProject
+{
+    public interface ICommand
+    {
+        void Execute();
+    }
+}
