+++ lab_06/ConsoleApp1/Program.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ConsoleApp1;
+using System;
 
 namespace ATMProject
 {
